Bring Green up to date with master

+++ src/server/job_func.c
@@ -142,12 +142,10 @@
 #include "portability.h"
 #include "array.h"
 #include "pbs_job.h"
-#include "resizable_array.h"
 #include "svr_func.h" /* get_svr_attr_* */
 #include "issue_request.h" /* release_req */
 #include "ji_mutex.h"
 #include "user_info.h"
-#include "svr_task.h"
 #include "mutex_mgr.hpp"
 #include "job_route.h" /* job_route */
 #include <string>
@@ -177,8 +175,8 @@ int issue_signal(job **, const char *, void(*)(batch_request *), void *, char *)
 static void job_init_wattr(job *);
 
 /* Global Data items */
-struct all_jobs        alljobs;
-extern struct all_jobs array_summary;
+all_jobs        alljobs;
+extern all_jobs array_summary;
 
 int check_job_log_started = 0;
 
@@ -492,7 +490,7 @@ int job_abt(
           if (depend_on_term(pjob) == PBSE_JOBNOTFOUND)
             {
             pjob = NULL;
-            pjob_mutex.set_lock_on_exit(false);
+            pjob_mutex.set_unlock_on_exit(false);
             }
           }
         
@@ -553,7 +551,7 @@ int job_abt(
       if (depend_on_term(pjob) == PBSE_JOBNOTFOUND)
         {
         pjob = NULL;
-        pjob_mutex.set_lock_on_exit(false);
+        pjob_mutex.set_unlock_on_exit(false);
         }
       /* pjob_mutex managed mutex already points to pjob->ji_mutex. Nothing to do */
       }
@@ -588,7 +586,7 @@ int job_abt(
     }
 
   if (pjob == NULL)
-    pjob_mutex.set_lock_on_exit(false);
+    pjob_mutex.set_unlock_on_exit(false);
 
   return(rc);
   }  /* END job_abt() */
@@ -777,6 +775,9 @@ void job_free(
    * the lock and then deletes the job, but thread 2 gets the job's lock as
    * the job is freed, causing segfaults. We use the recycler and the 
    * ji_being_recycled flag to solve this problem --dbeer */
+
+  remove_job(&alljobs,pj); //Remove this from the alljobs array.
+
   if (use_recycle)
     {
     insert_into_recycler(pj);
@@ -853,7 +854,7 @@ void job_free(
  * job_clone - create a clone of a job for use with job arrays
  */
 
-/*static*/ job *job_clone(
+job *job_clone(
 
   job       *template_job, /* I */  /* job to clone */
   job_array *pa,           /* I */  /* array which the job is a part of */
@@ -880,7 +881,7 @@ void job_free(
 
   if (LOGLEVEL >= 7)
     {
-    sprintf(log_buf, "taskid %d", taskid);
+    sprintf(log_buf, "taskid %d jobid %s ", taskid, template_job->ji_qs.ji_jobid);
     log_event(PBSEVENT_JOB, PBS_EVENTCLASS_JOB, __func__, log_buf);
     }
 
@@ -893,8 +894,8 @@ void job_free(
 
   if (template_job == NULL)
     {
-      log_err(PBSE_BAD_PARAMETER, __func__, "template_job* is NULL");
-      return(NULL);
+    log_err(PBSE_BAD_PARAMETER, __func__, "template_job* is NULL");
+    return(NULL);
     }
 
   if ((pnewjob = job_alloc()) == NULL)
@@ -1116,7 +1117,7 @@ void *job_clone_wt(
   job                *pjobclone;
   char               *jobid;
   int                 i;
-  int                 prev_index = -1;
+  char                *prev_job_id = NULL;
   int                 actual_job_count = 0;
   int                 newstate;
   int                 newsub;
@@ -1193,7 +1194,8 @@ void *job_clone_wt(
         }
 
       mutex_mgr clone_mgr(pjobclone->ji_mutex, true);
-      svr_evaljobstate(pjobclone, &newstate, &newsub, 1);
+
+      svr_evaljobstate(*pjobclone, newstate, newsub, 1);
 
       /* do this so that  svr_setjobstate() doesn't alter sv_jobstates,
        * these are set later in svr_enquejob() */
@@ -1207,10 +1209,10 @@ void *job_clone_wt(
 
       array_mgr.unlock();
 
-      if ((rc = svr_enquejob(pjobclone, FALSE, prev_index, false)))
+      if ((rc = svr_enquejob(pjobclone, FALSE, prev_job_id, false)))
         {
         /* XXX need more robust error handling */
-        clone_mgr.set_lock_on_exit(false);
+        clone_mgr.set_unlock_on_exit(false);
 
         if (rc != PBSE_JOB_RECYCLED)
           {
@@ -1218,7 +1220,10 @@ void *job_clone_wt(
           }
 
         if ((pa = get_array(arrayid)) == NULL)
+          {
+          if(prev_job_id != NULL) free(prev_job_id);
           return(NULL);
+          }
 
         array_mgr.mark_as_locked();
 
@@ -1228,8 +1233,12 @@ void *job_clone_wt(
       if ((pa = get_jobs_array(&pjobclone)) == NULL)
         {
         if (pjobclone == NULL)
-          clone_mgr.set_lock_on_exit(false);
+          {
+          /* pjobclone has been released. No mutex left to unlock */
+          clone_mgr.set_unlock_on_exit(false);
+          }
 
+        if(prev_job_id != NULL) free(prev_job_id);
         return(NULL);
         }
       
@@ -1242,22 +1251,30 @@ void *job_clone_wt(
         svr_job_purge(pjobclone);
         
         if ((pa = get_array(arrayid)) == NULL)
+          {
+          if(prev_job_id != NULL) free(prev_job_id);
           return(NULL);
+          }
 
         array_mgr.mark_as_locked();
         
         continue;
         }
       
-      prev_index = get_jobs_index(&alljobs, pjobclone);
+      if(prev_job_id != NULL) free(prev_job_id);
+      prev_job_id = NULL;
+      if(alljobs.find(pjobclone->ji_qs.ji_jobid) != NULL)
+        {
+        prev_job_id = strdup(pjobclone->ji_qs.ji_jobid);
+        }
       
       pa->ai_qs.num_cloned++;
       
       rn->start++;
       
       /* index below 0 means the job no longer exists */
-      if (prev_index < 0)
-        clone_mgr.set_lock_on_exit(false);
+      if (prev_job_id == NULL)
+        clone_mgr.set_unlock_on_exit(false);
       }  /* END for (i) */
 
     if (rn->start > rn->end)
@@ -1267,6 +1284,9 @@ void *job_clone_wt(
       }
     }    /* END while (loop) */
       
+  if(prev_job_id != NULL) free(prev_job_id);
+  prev_job_id = NULL;
+
   array_save(pa);
 
   /* scan over all the jobs in the array and unset the hold */
@@ -1312,7 +1332,7 @@ void *job_clone_wt(
         }
       
       pjob->ji_modified = TRUE;
-      svr_evaljobstate(pjob, &newstate, &newsub, 1);
+      svr_evaljobstate(*pjob, newstate, newsub, 1);
       svr_setjobstate(pjob, newstate, newsub, FALSE);
 
        /*
@@ -1795,7 +1815,8 @@ int record_jobinfo(
 
 int svr_job_purge(
 
-  job *pjob)  /* I (modified) */
+  job *pjob,  /* I (modified) */
+  int leaveSpoolFiles) /* I */
 
   {
   int           rc = PBSE_NONE;
@@ -1851,8 +1872,8 @@ int svr_job_purge(
   if ((job_has_arraystruct == FALSE) || (job_is_array_template == TRUE))
     if (remove_job(&array_summary,pjob) == PBSE_JOB_RECYCLED)
       {
-      /* PBSE_JOB_RECYCLED means the job is gone */
-      pjob_mutex.set_lock_on_exit(false);
+      /* PBSE_JOB_RECYCLED means the job is gone. remove_job alreadly unlocked pjob->ji_mutex */
+      pjob_mutex.set_unlock_on_exit(false); 
       return(PBSE_NONE);
       }
 
@@ -1894,7 +1915,7 @@ int svr_job_purge(
       }
     else
       {
-      pjob_mutex.set_lock_on_exit(false);
+      pjob_mutex.set_unlock_on_exit(false);
       return(PBSE_JOBNOTFOUND);
       }
     }
@@ -1921,7 +1942,7 @@ int svr_job_purge(
       if (pjob->ji_being_recycled == FALSE)
         {
         job_free(pjob, TRUE);
-        pjob_mutex.set_lock_on_exit(false);
+        pjob_mutex.set_unlock_on_exit(false);  /* job_free will release lock */
         }
       else
         pjob_mutex.unlock();
@@ -1930,9 +1951,10 @@ int svr_job_purge(
   else
     {
     job_free(pjob, TRUE);
-    pjob_mutex.set_lock_on_exit(false);
+    pjob_mutex.set_unlock_on_exit(false); /* job_free will release lock */
     }
 
+  /* pjob->ji_mutex is unlocked at this point */
   /* delete the script file */
   if ((job_has_arraystruct == FALSE) || 
       (job_is_array_template == TRUE))
@@ -1953,34 +1975,37 @@ int svr_job_purge(
       }
     }
 
-  /* delete any spooled stdout */
-  snprintf(namebuf, sizeof(namebuf), "%s%s%s", path_jobs, job_fileprefix, JOB_STDOUT_SUFFIX);
-
-  if (unlink(namebuf) < 0)
+  if (!leaveSpoolFiles)
     {
-    if (errno != ENOENT)
-      log_err(errno, __func__, msg_err_purgejob);
-    }
-  else if (LOGLEVEL >= 6)
-    {
-    sprintf(log_buf, "removed job stdout");
+    /* delete any spooled stdout */
+    snprintf(namebuf, sizeof(namebuf), "%s%s%s", path_spool, job_fileprefix, JOB_STDOUT_SUFFIX);
 
-    log_record(PBSEVENT_DEBUG, PBS_EVENTCLASS_JOB, job_id, log_buf);
-    }
+    if (unlink(namebuf) < 0)
+      {
+      if (errno != ENOENT)
+        log_err(errno, __func__, msg_err_purgejob);
+      }
+    else if (LOGLEVEL >= 6)
+      {
+      sprintf(log_buf, "removed job stdout");
 
-  /* delete any spooled stderr */
-  snprintf(namebuf, sizeof(namebuf), "%s%s%s", path_jobs, job_fileprefix, JOB_STDERR_SUFFIX);
+      log_record(PBSEVENT_DEBUG, PBS_EVENTCLASS_JOB, job_id, log_buf);
+      }
 
-  if (unlink(namebuf) < 0)
-    {
-    if (errno != ENOENT)
-      log_err(errno, __func__, msg_err_purgejob);
-    }
-  else if (LOGLEVEL >= 6)
-    {
-    sprintf(log_buf, "removed job stderr");
+    /* delete any spooled stderr */
+    snprintf(namebuf, sizeof(namebuf), "%s%s%s", path_spool, job_fileprefix, JOB_STDERR_SUFFIX);
 
-    log_record(PBSEVENT_DEBUG, PBS_EVENTCLASS_JOB, job_id, log_buf);
+    if (unlink(namebuf) < 0)
+      {
+      if (errno != ENOENT)
+        log_err(errno, __func__, msg_err_purgejob);
+      }
+    else if (LOGLEVEL >= 6)
+      {
+      sprintf(log_buf, "removed job stderr");
+
+      log_record(PBSEVENT_DEBUG, PBS_EVENTCLASS_JOB, job_id, log_buf);
+      }
     }
 
   /* delete checkpoint file directory if there is one */
@@ -2074,7 +2099,7 @@ job_array *get_jobs_array(
     }
 
   mutex_mgr job_mutex(pjob->ji_mutex,true);
-  job_mutex.set_lock_on_exit(false);
+  job_mutex.set_unlock_on_exit(false);
 
   strcpy(jobid, pjob->ji_qs.ji_jobid);
